Companion Hunger
The companion now has a hunger metre that fills the more you feed him.
The hunger metre drops by 1 (ATM) everysecond

+++ Match3Game/Assets/Scripts/DotManagerScript.cs
@@ -28,7 +28,7 @@ public class DotManagerScript : MonoBehaviour
     public int BlueScore;
     public int YellowScore;
     public int GreenScore;
-    public int TotalScore;
+    public float TotalScore;
     public int Limit;
     public int NumOfPeices;
     public Text HighScore;
